+++ Build-Release.ps1
@@ -1,7 +1,7 @@
 [CmdletBinding()]
 param(
     [ValidateNotNullOrEmpty()]
-    [string]$Version = '1.3.1-dev.3.1',
+    [string]$Version = '1.3.1-dev.4',
 
     [ValidateNotNullOrEmpty()]
     [string]$Runtime = 'win-x64',